requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pycryptodome==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
